TIP-1287: move RemoveCategoryFilterInJobInstanceSubscriber
