+++ src/Akeneo/Pim/Enrichment/Bundle/Resources/config/event_subscribers.yml
@@ -117,7 +117,7 @@ services:
           - { name: kernel.event_subscriber }
 
     pim_enrich.event_listener.remove_category_filter_in_job_instance:
-        class: Akeneo\Pim\Enrichment\Bundle\EventSubscriber\Storage\RemoveCategoryFilterInJobInstanceSubscriber
+        class: Akeneo\Pim\Enrichment\Bundle\EventSubscriber\Category\OnDelete\RemoveCategoryFilterInJobInstanceSubscriber
         arguments:
           - '@akeneo_batch.job.job_instance_repository'
           - '@akeneo_batch.saver.job_instance'
