+++ tests/back/Pim/Enrichment/Specification/Bundle/EventSubscriber/Category/OnDelete/RemoveCategoryFilterInJobInstanceSubscriberSpec.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Specification\Akeneo\Pim\Enrichment\Bundle\EventSubscriber\Storage;
+namespace Specification\Akeneo\Pim\Enrichment\Bundle\EventSubscriber\Category\OnDelete;
 
-use Akeneo\Pim\Enrichment\Bundle\EventSubscriber\Storage\RemoveCategoryFilterInJobInstanceSubscriber;
+use Akeneo\Pim\Enrichment\Bundle\EventSubscriber\Category\OnDelete\RemoveCategoryFilterInJobInstanceSubscriber;
 use Akeneo\Pim\Enrichment\Component\Category\Model\Category;
 use Akeneo\Tool\Component\Batch\Model\JobInstance;
 use Akeneo\Tool\Component\StorageUtils\Saver\BulkSaverInterface;
